+++ kglib/kgcn_experimental/examples/genealogy/data_test.py
@@ -19,7 +19,7 @@
 from kglib.kgcn.core.ingest.traverse.data.context.neighbour import build_thing
 import networkx as nx
 
-from kglib.kgcn_experimental.genealogy.data import create_concept_graphs
+from kglib.kgcn_experimental.examples.genealogy.data import create_concept_graphs
 from kglib.utils.graph.test.case import GraphTestCase
 from grakn.client import GraknClient
