+++ kglib/kgcn_experimental/examples/genealogy/genealogy_training.gql
