+++ kglib/kgcn_experimental/examples/genealogy/in_memory/data.py
